backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
strawberry-graphql[fastapi]>=0.219.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
Faker
sqlalchemy
psycopg2-binary
python-dotenv